update | fix | add update: UI fix: Overflow bug and numericTextBox add: feature Login for "add to bag" and "buy now button'
