+++ WPFEcommerceApp/WPFEcommerceApp/Screens/General/ProductDetail/ProductDetailBanned/ProductDetailBanned.xaml
@@ -1,0 +1,51 @@
+﻿<UserControl x:Class="WPFEcommerceApp.ProductDetailBanned"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:WPFEcommerceApp"
+             xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
+             xmlns:c="clr-namespace:CalcBinding;assembly=CalcBinding"
+             xmlns:i="http://schemas.microsoft.com/xaml/behaviors"
+             mc:Ignorable="d">
+    <Grid
+        Background="{StaticResource BoxShadowColorBrush}">
+        <StackPanel
+                VerticalAlignment="Center"
+                HorizontalAlignment="Center">
+            <materialDesign:PackIcon
+                HorizontalAlignment="Center"
+                Kind="DoNotEnter"
+                Height="100"
+                Width="100"/>
+            <StackPanel
+                Margin="0 30 0 0"
+                Orientation="Horizontal"
+                HorizontalAlignment="Center">
+                <TextBlock
+                    VerticalAlignment="Center"
+                    Text="The product has been banned from us."
+                    FontSize="{StaticResource FontSizeText}"
+                    FontFamily="{StaticResource Medium}"/>
+                <TextBlock
+                    VerticalAlignment="Center"
+                    Text=" "
+                    FontSize="{StaticResource FontSizeText}"
+                    FontFamily="{StaticResource Medium}"/>
+                <Label
+                    x:Name="linkLabel"
+                    FontSize="{StaticResource FontSizeText}"
+                    FontFamily="{StaticResource Medium}">
+                    <Hyperlink
+                        Foreground="#4B8CDE"
+                        Command="{Binding ContactCommand,UpdateSourceTrigger=PropertyChanged}"
+                        CommandParameter="{Binding ElementName=linkLabel, Path=IsEnabled}">
+                        <TextBlock
+                            Text="Contact us">
+                        </TextBlock>
+                    </Hyperlink>
+                </Label>
+            </StackPanel>
+        </StackPanel>
+    </Grid>
+</UserControl>
